4.23

+++ osa4/blogilista/tests/blog_api.test.js
@@ -4,18 +4,24 @@ const app = require("../app");
 const api = supertest(app);
 const Blog = require("../models/blog");
 const helper = require("./test_helper");
-
+const User = require("../models/user");
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+let token = null;
 beforeEach(async () => {
   await Blog.deleteMany({});
-  let blogObject = new Blog(helper.initialBlogs[0]);
-  await blogObject.save();
-  blogObject = new Blog(helper.initialBlogs[1]);
-  await blogObject.save();
+  await Blog.insertMany(helper.initialBlogs);
+  await User.deleteMany({});
+  const passwordHash = await bcrypt.hash("hunter2", 10);
+  const user = await new User({ username: "name", passwordHash }).save();
+  const userForToken = { username: "name", id: user.id };
+  return (token = jwt.sign(userForToken, process.env.SECRET));
 });
 
 test("blogs are returned as json", async () => {
   await api
     .get("/api/blogs")
+    .set("Authorization", `Bearer ${token}`)
     .expect(200)
     .expect("Content-Type", /application\/json/);
 });
@@ -23,6 +29,7 @@ test("blogs are returned as json", async () => {
 test("blogs have id", async () => {
   const response = await api
     .get("/api/blogs")
+    .set("Authorization", `Bearer ${token}`)
     .expect(200)
     .expect("Content-Type", /application\/json/);
 
@@ -31,13 +38,13 @@ test("blogs have id", async () => {
   }
 });
 test("there are two blogs", async () => {
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
 
   expect(response.body).toHaveLength(helper.initialBlogs.length);
 });
 
 test("a specific blog is within the returned blogs", async () => {
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
 
   const titles = response.body.map((r) => r.title);
   expect(titles).toContain("testing");
@@ -53,11 +60,12 @@ test("a valid blog can be added ", async () => {
 
   await api
     .post("/api/blogs")
+    .set("Authorization", `Bearer ${token}`)
     .send(newBlog)
     .expect(201)
     .expect("Content-Type", /application\/json/);
 
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
 
   const contents = response.body.map((r) => r.title);
 
@@ -72,9 +80,9 @@ test("blog without likes has 0 likes", async () => {
     url: "nolikes.com",
   };
 
-  await api.post("/api/blogs").send(newBlog).expect(201);
+  await api.post("/api/blogs").set("Authorization", `Bearer ${token}`).send(newBlog).expect(201);
 
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
 
   for (const blog of response.body) {
     if (blog.title === "nolikes") {
@@ -90,9 +98,9 @@ test("blog without title", async () => {
     likes: 5,
   };
 
-  await api.post("/api/blogs").send(newBlog).expect(400);
+  await api.post("/api/blogs").set("Authorization", `Bearer ${token}`).send(newBlog).expect(400);
 
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
   expect(response.body).toHaveLength(helper.initialBlogs.length);
 });
 
@@ -103,21 +111,21 @@ test("blog without url", async () => {
     likes: 5,
   };
 
-  await api.post("/api/blogs").send(newBlog).expect(400);
+  await api.post("/api/blogs").set("Authorization", `Bearer ${token}`).send(newBlog).expect(400);
 
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
   expect(response.body).toHaveLength(helper.initialBlogs.length);
 });
 
 test("removing a single blog works", async () => {
-  const response = await api.get("/api/blogs");
-  await api.delete(`/api/blogs/${response.body[0].id}`).expect(204);
-  const after = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
+  await api.delete(`/api/blogs/${response.body[0].id}`).set("Authorization", `Bearer ${token}`).expect(204);
+  const after = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
   expect(after.body).toHaveLength(response.body.length - 1);
 });
 
 test("updating a blog works", async () => {
-  const response = await api.get("/api/blogs");
+  const response = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
   const updatedBlog = {
     title: "updatedTitle",
     author: "updatedAuthor",
@@ -127,9 +135,10 @@ test("updating a blog works", async () => {
   };
   await api
     .put(`/api/blogs/${response.body[0].id}`)
+    .set("Authorization", `Bearer ${token}`)
     .send(updatedBlog)
     .expect(200);
-  const after = await api.get("/api/blogs");
+  const after = await api.get("/api/blogs").set("Authorization", `Bearer ${token}`);
   expect(after.body).toContainEqual(updatedBlog);
 });
 afterAll(() => {
